Renamed module to allow more logical importing
The import methods now can look testobject.devices inside the client file.
